refactor(test): move tests closer to the source code, clean up eslint startup parameters
