+++ .eslintrc.js
@@ -41,12 +41,15 @@ module.exports = {
   },
   overrides: [
     {
-      files: ['**/*.ts', '**/*.tsx'],
-    },
-    {
-      files: '*.spec.js',
+      files: '*.spec.ts',
       rules: {
-        'no-unused-expressions': 'off',
+        'no-undef': 0,
+        'no-unused-expressions': 0,
+        // I want to test weird cases
+        '@typescript-eslint/ban-ts-comment': 0,
+        'import/no-extraneous-dependencies': 0,
+        '@typescript-eslint/no-explicit-any': 0,
+        '@typescript-eslint/no-non-null-assertion': 0,
       },
     },
   ],
