--- tests/.eslintrc.js
@@ -1,11 +1,0 @@
-module.exports = {
-  rules: {
-    'no-undef': 0,
-    'no-unused-expressions': 0,
-    // I want to test weird cases
-    '@typescript-eslint/ban-ts-comment': 0,
-    'import/no-extraneous-dependencies': 0,
-    '@typescript-eslint/no-explicit-any': 0,
-    '@typescript-eslint/no-non-null-assertion': 0,
-  },
-};
